refactor(utils): move isNode to element
